+++ HISTORY.rst
@@ -2,11 +2,26 @@
 History
 =======
 
-Unlreleased / main
-------------------
+Unreleased
+----------
+* Use ``argparse`` for all command line tools and make use of ``workflows`` transport 
+  argument injection. Minimum ``workflows`` version is now 2.14
 * Add a zocalo plugin for ``ispyb.simulate`` 
-  (`#163 <https://github.com/DiamondLightSource/ispyb-api/pull/163>`_) to 
-  call a recipe before and after creation of data
+  (`#163 <https://github.com/DiamondLightSource/ispyb-api/pull/163>`_) to call a recipe
+  before and after creation of data
+
+0.10.0 (2021-10-04)
+-------------------
+* New ``zocalo.shutdown`` command to shutdown Zocalo services
+* New ``zocalo.queue_drain`` command to drain one queue into another in a controlled manner
+* New ``zocalo.util.rabbitmq.http_api_request()`` utility function to return a
+  ``urllib.request.Request`` object to query the RabbitMQ API using the credentials
+  specified via ``zocalo.configuration``.
+* ``zocalo.wrap`` now emits tracebacks on hard crashes and ``SIGUSR2`` signals
+
+0.9.1 (2021-08-18)
+------------------
+* Expand ~ in paths in configuration files
 
 0.9.0 (2021-08-18)
 ------------------